Dependencies updated
The dependencies in requirements.text have been updated. Due to a
problem with PyShExC library, wikibase-sync cannot be run with a Python
version above 3.9

Some tests are not passing because they needed a Wikibase instance that
is no longer available. Even though there is a new instance, already
referenced in the code, it needs
some information to be already present for the tests to pass.